Fix procedure (#672)
* modify procedure

* fix procedure

* fix procedure

* fix procedure

* fix procedure

* fix procedure
