+++ procedures/custom_cpp/dc_procedure.cpp
@@ -14,40 +14,19 @@
 
 #include "lgraph/olap_on_db.h"
 #include "tools/json.hpp"
-#include "./algo.h"
+#include "../algo_cpp/algo.h"
 
 using namespace lgraph_api;
 using namespace lgraph_api::olap;
 using json = nlohmann::json;
-
-void CountComp(OlapBase<Empty>& graph, ParallelVector<size_t>& label, size_t& max, size_t& num) {
-    ParallelVector<size_t> cnt = graph.AllocVertexArray<size_t>();
-    cnt.Fill(0);
-    graph.ProcessVertexInRange<size_t>([&](size_t v) {
-      if (graph.OutDegree(v) == 0 && graph.InDegree(v) == 0) return 0;
-      size_t v_label = label[v];
-      write_add(&cnt[v_label], (size_t)1);
-      return 0;
-    }, 0, label.Size());
-    max = 0;
-    num = graph.ProcessVertexInRange<size_t>([&](size_t v) {
-      if (max < cnt[v])
-          write_max(&max, cnt[v]);
-      return ((graph.OutDegree(v) > 0 || graph.InDegree(v) > 0) && cnt[v] > 0) ? 1 : 0;
-    }, 0, label.Size());
-}
-
 
 extern "C" bool Process(GraphDB& db, const std::string& request, std::string& response) {
     double start_time;
 
     // prepare
     start_time = get_time();
-    std::string output_file = "";
-    std::cout << "Input: " << request << std::endl;
     try {
         json input = json::parse(request);
-        parse_from_json(output_file, "output_file", input);
     } catch (std::exception& e) {
         response = "json parse error: " + std::string(e.what());
         std::cout << response << std::endl;
@@ -59,30 +38,18 @@ extern "C" bool Process(GraphDB& db, const std::string& request, std::string& re
 
     // core
     start_time = get_time();
-    auto label = olapondb.AllocVertexArray<size_t>();
-    SCCCore(olapondb, label);
-    size_t num_components, max_component;
-    CountComp(olapondb, label, max_component, num_components);
+    double total_dc = DCCore(olapondb);
     auto core_cost = get_time() - start_time;
-
-    // output
-    start_time = get_time();
-    if (output_file != "") {
-        olapondb.WriteToFile<size_t>(true, label, output_file);
-    }
-    auto output_cost = get_time() - start_time;
 
     // return
     {
         json output;
-        output["num_components"] = num_components;
-        output["max_component"] = max_component;
+        output["graph_dc"] = total_dc;
         output["num_vertices"] = olapondb.NumVertices();
         output["num_edges"] = olapondb.NumEdges();
         output["prepare_cost"] = prepare_cost;
         output["core_cost"] = core_cost;
-        output["output_cost"] = output_cost;
-        output["total_cost"] = prepare_cost + core_cost + output_cost;
+        output["total_cost"] = prepare_cost + core_cost;
         response = output.dump();
     }
     return true;
