+++ procedures/custom_cpp/leiden_procedure.cpp
@@ -1,4 +1,4 @@
-﻿/**
+/**
  * Copyright 2022 AntGroup CO., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,7 +14,7 @@
 
 #include "lgraph/olap_on_db.h"
 #include "tools/json.hpp"
-#include "./algo.h"
+#include "../algo_cpp/algo.h"
 
 using namespace lgraph_api;
 using namespace lgraph_api::olap;
@@ -25,52 +25,61 @@ extern "C" bool Process(GraphDB& db, const std::string& request, std::string& re
 
     // prepare
     start_time = get_time();
-    std::string root_value = "0";
-    std::string root_label = "node";
-    std::string root_field = "id";
+    double gamma = 0.2;
+    double theta = 0.1;
+    unsigned random_seed = 0;
+    size_t threshold = 0;
+    std::string weight = "";
     std::string output_file = "";
     std::cout << "Input: " << request << std::endl;
     try {
         json input = json::parse(request);
-        parse_from_json(root_value, "root_value", input);
-        parse_from_json(root_label, "root_label", input);
-        parse_from_json(root_field, "root_field", input);
+        parse_from_json(gamma, "gamma", input);
+        assert(gamma > 0 && gamma <= 1);
+        parse_from_json(theta, "theta", input);
+        assert(theta > 0 && theta <= 1);
+        parse_from_json(random_seed, "random_seed", input);
+        parse_from_json(weight, "weight", input);
+        parse_from_json(threshold, "threshold", input);
         parse_from_json(output_file, "output_file", input);
     } catch (std::exception& e) {
         response = "json parse error: " + std::string(e.what());
         std::cout << response << std::endl;
         return false;
     }
-
     auto txn = db.CreateReadTxn();
-    int64_t root_vid =
-        txn.GetVertexIndexIterator(root_label, root_field, root_value, root_value).GetVid();
-    OlapOnDB<Empty> olapondb(db, txn, SNAPSHOT_PARALLEL);
+    auto edge_convert = [&](OutEdgeIterator& eit, double& edge_data) -> bool {
+        if (weight.length() == 0) {
+            edge_data = 1;
+            return true;
+        }
+        edge_data = eit.GetField(weight).real();
+        if (edge_data > 0.0) return true;
+        return false;
+    };
+    OlapOnDB<double> olapondb(db, txn, SNAPSHOT_PARALLEL | SNAPSHOT_UNDIRECTED, nullptr,
+                              edge_convert);
+    printf("|V| = %lu\n", olapondb.NumVertices());
+    printf("|E| = %lu\n", olapondb.NumEdges());
     auto prepare_cost = get_time() - start_time;
 
     // core
     start_time = get_time();
-    ParallelVector<size_t> parent = olapondb.AllocVertexArray<size_t>();
-    size_t count = BFSCore(olapondb, olapondb.MappedVid(root_vid), parent);
-    printf("found_vertices = %ld\n", count);
+    ParallelVector<size_t> label = olapondb.AllocVertexArray<size_t>();
+    LeidenCore(olapondb, label, random_seed, theta, gamma, threshold);
     auto core_cost = get_time() - start_time;
 
     // output
     start_time = get_time();
-    // TODO(any): write parent back to graph
     if (output_file != "") {
-        olapondb.WriteToFile<size_t>(true, parent, output_file,
-                                    [&](size_t vid, size_t vdata) -> bool {
-            return vdata != (size_t)-1;
-        });
+        olapondb.WriteToFile<size_t>(true, label, output_file);
     }
 
-    auto output_cost = get_time() - start_time;
+    double output_cost = get_time() - start_time;
 
     // return
     {
         json output;
-        output["found_vertices"] = count;
         output["num_vertices"] = olapondb.NumVertices();
         output["num_edges"] = olapondb.NumEdges();
         output["prepare_cost"] = prepare_cost;
