+++ procedures/custom_cpp/lpa_procedure.cpp
@@ -1,4 +1,4 @@
-﻿/**
+/**
  * Copyright 2022 AntGroup CO., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,7 +14,7 @@
 
 #include "lgraph/olap_on_db.h"
 #include "tools/json.hpp"
-#include "./algo.h"
+#include "../algo_cpp/algo.h"
 
 using namespace lgraph_api;
 using namespace lgraph_api::olap;
@@ -26,15 +26,13 @@ using json = nlohmann::json;
  * @param db The reference to the GraphDB object on which the request will be processed.
  * @param request The input request in JSON format.
  *        The request should contain the following parameters:
- *        - "num_iterations": The number of iterations.
  *        - "vertex_label_filter": Filter vertex sets based on vertex labels
  *        - "edge_label_filter": Filter edge sets based on edge labels
- *        - "pr_value": Vertex field name to be written back into the database.
- *        - "output_file": Pr value to be written to the file.
+ *        - "lpa_value": Vertex field name to be written back into the database.
+ *        - "output_file": Lpa value to be written to the file.
  * @param response The output response in JSON format.
  *       The response will contain the following parameters:
- *       - "max_pr_id": The vertex id with the highest pagerank value.
- *       - "max_pr_val": The highest pagerank value.
+ *       - "modularity": The modularity of the graph.
  *       - "num_vertices": The number of vertices in the graph.
  *       - "num_edges": The number of edges in the graph.
  *       - "prepare_cost": The time cost of preparing the graph data.
@@ -50,24 +48,24 @@ extern "C" bool Process(GraphDB& db, const std::string& request, std::string& re
     // prepare
     start_time = get_time();
     int num_iterations = 20;
+    int sync_flag = 1;
     std::string vertex_label_filter = "";
     std::string edge_label_filter = "";
-    std::string pr_value = "";
+    std::string lpa_value = "";
     std::string output_file = "";
-    std::cout << "Input: " << request << std::endl;
     try {
         json input = json::parse(request);
         parse_from_json(num_iterations, "num_iterations", input);
+        parse_from_json(sync_flag, "sync_flag", input);
         parse_from_json(vertex_label_filter, "vertex_label_filter", input);
         parse_from_json(edge_label_filter, "edge_label_filter", input);
-        parse_from_json(pr_value, "pr_value", input);
+        parse_from_json(lpa_value, "lpa_value", input);
         parse_from_json(output_file, "output_file", input);
     } catch (std::exception& e) {
         response = "json parse error: " + std::string(e.what());
         std::cout << response << std::endl;
         return false;
     }
-
     auto txn = db.CreateReadTxn();
     std::function<bool(VertexIterator &)> vertex_filter = nullptr;
     std::function<bool(OutEdgeIterator &, Empty &)> edge_filter = nullptr;
@@ -82,47 +80,32 @@ extern "C" bool Process(GraphDB& db, const std::string& request, std::string& re
             return eit.GetLabel() == edge_label_filter;
         };
     }
-    OlapOnDB<Empty> olapondb(db, txn, SNAPSHOT_PARALLEL, vertex_filter, edge_filter);
+    OlapOnDB<Empty> olapondb(db, txn, SNAPSHOT_PARALLEL | SNAPSHOT_UNDIRECTED,
+            vertex_filter, edge_filter);
     auto prepare_cost = get_time() - start_time;
 
     // core
     start_time = get_time();
-    ParallelVector<double> pr = olapondb.AllocVertexArray<double>();
-    PageRankCore(olapondb, num_iterations, pr);
-    auto all_vertices = olapondb.AllocVertexSubset();
-    all_vertices.Fill();
-    size_t max_pr_vi = olapondb.ProcessVertexActive<size_t>(
-        [&](size_t vi) { return vi; }, all_vertices, 0,
-        [&](size_t a, size_t b) { return pr[a] > pr[b] ? a : b; });
+    ParallelVector<size_t> label = olapondb.AllocVertexArray<size_t>();
+    double modularity = LPACore(olapondb, label, num_iterations, sync_flag);
     auto core_cost = get_time() - start_time;
 
     // output
     start_time = get_time();
     if (output_file != "") {
-        olapondb.WriteToFile<double>(true, pr, output_file);
+        olapondb.WriteToFile<size_t>(true, label, output_file);
     }
     txn.Commit();
 
-    if (pr_value != "") {
-        olapondb.WriteToGraphDB<double>(pr, pr_value);
+    if (lpa_value != "") {
+        olapondb.WriteToGraphDB<size_t>(label, lpa_value);
     }
-
-    printf("max rank value is pr[%ld] = %lf\n", olapondb.OriginalVid(max_pr_vi), pr[max_pr_vi]);
     auto output_cost = get_time() - start_time;
 
-
-    auto vit = txn.GetVertexIterator(olapondb.OriginalVid(max_pr_vi), false);
-    auto vit_label = vit.GetLabel();
-    auto primary_field = txn.GetVertexPrimaryField(vit_label);
-    auto field_data = vit.GetField(primary_field);
     // return
     {
         json output;
-        output["max_pr_id"] = olapondb.OriginalVid(max_pr_vi);
-        output["max_pr_label"] = vit_label;
-        output["max_pr_primaryfield"] = primary_field;
-        output["max_pr_fielddata"] = field_data.ToString();
-        output["max_pr_val"] = pr[max_pr_vi];
+        output["modularity"] = modularity;
         output["num_vertices"] = olapondb.NumVertices();
         output["num_edges"] = olapondb.NumEdges();
         output["prepare_cost"] = prepare_cost;
