+++ procedures/custom_cpp/sssp_procedure.cpp
@@ -1,4 +1,4 @@
-﻿/**
+/**
  * Copyright 2022 AntGroup CO., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,31 +14,11 @@
 
 #include "lgraph/olap_on_db.h"
 #include "tools/json.hpp"
-#include "./algo.h"
+#include "../algo_cpp/algo.h"
 
 using namespace lgraph_api;
 using namespace lgraph_api::olap;
 using json = nlohmann::json;
-
-void CountComp(OlapBase<Empty>& graph, ParallelVector<size_t>& label, size_t& max, size_t& num) {
-    ParallelVector<size_t> cnt = graph.AllocVertexArray<size_t>();
-    cnt.Fill(0);
-    graph.ProcessVertexInRange<size_t>(
-        [&](size_t v) {
-            if (graph.OutDegree(v) == 0) return 0;
-            size_t v_label = label[v];
-            write_add(&cnt[v_label], (size_t)1);
-            return 0;
-        },
-        0, label.Size());
-    max = 0;
-    num = graph.ProcessVertexInRange<size_t>(
-        [&](size_t v) {
-            if (max < cnt[v]) write_max(&max, cnt[v]);
-            return (graph.OutDegree(v) > 0 && cnt[v] > 0) ? 1 : 0;
-        },
-        0, label.Size());
-}
 
 /**
  * Processes a request on a GraphDB.
@@ -46,14 +26,14 @@ void CountComp(OlapBase<Empty>& graph, ParallelVector<size_t>& label, size_t& ma
  * @param db The reference to the GraphDB object on which the request will be processed.
  * @param request The input request in JSON format.
  *        The request should contain the following parameters:
- *        - "vertex_label_filter": Filter vertex sets based on vertex labels
- *        - "edge_label_filter": Filter edge sets based on edge labels
- *        - "wcc_value": Vertex field name to be written back into the database.
- *        - "output_file": Wcc value to be written to the file.
+ *        - "root_value": The value of the root vertex.
+ *        - "root_label": The label of the root vertex.
+ *        - "root_field": The field of the root vertex.
+ *        - "output_file": Sssp distance to be written to the file.
  * @param response The output response in JSON format.
  *        The response will contain the following parameters:
- *        - "num_components": The number of components in the graph.
- *        - "max_component": The size of the largest component.
+ *        - "max_distance_vid": The vertex id with the maximum distance.
+ *        - "max_distance_val": The maximum distance.
  *        - "num_vertices": The number of vertices in the graph.
  *        - "num_edges": The number of edges in the graph.
  *        - "prepare_cost": The time cost of preparing the graph data.
@@ -62,50 +42,83 @@ void CountComp(OlapBase<Empty>& graph, ParallelVector<size_t>& label, size_t& ma
  *        - "total_cost": The total time cost.
  * @return True if the request is processed successfully, false otherwise.
  */
-
 extern "C" bool Process(GraphDB& db, const std::string& request, std::string& response) {
     double start_time;
 
     // prepare
     start_time = get_time();
+    std::string root_value = "0";
+    std::string root_label = "node";
+    std::string root_field = "id";
     std::string output_file = "";
     std::cout << "Input: " << request << std::endl;
     try {
         json input = json::parse(request);
+        parse_from_json(root_value, "root_value", input);
+        parse_from_json(root_label, "root_label", input);
+        parse_from_json(root_field, "root_field", input);
         parse_from_json(output_file, "output_file", input);
     } catch (std::exception& e) {
         response = "json parse error: " + std::string(e.what());
         std::cout << response << std::endl;
         return false;
     }
-
     auto txn = db.CreateReadTxn();
-    OlapOnDB<Empty> olapondb(db, txn, SNAPSHOT_PARALLEL | SNAPSHOT_UNDIRECTED);
+
+    OlapOnDB<double> olapondb(db, txn, SNAPSHOT_PARALLEL);
+    int64_t root_vid =
+        txn.GetVertexIndexIterator(root_label, root_field, root_value, root_value).GetVid();
     auto prepare_cost = get_time() - start_time;
 
     // core
     start_time = get_time();
-    ParallelVector<size_t> label = olapondb.AllocVertexArray<size_t>();
-    WCCCore(olapondb, label);
+    ParallelVector<double> distance = olapondb.AllocVertexArray<double>();
+    SSSPCore(olapondb, olapondb.MappedVid(root_vid), distance);
     auto core_cost = get_time() - start_time;
 
     // output
     start_time = get_time();
     if (output_file != "") {
-        olapondb.WriteToFile<size_t>(true, label, output_file);
+        olapondb.WriteToFile<double>(true, distance, output_file,
+                                    [&](size_t vid, double& vdata) -> bool {
+            return vdata != SSSP_INIT_VALUE;
+        });
     }
+    txn.Commit();
 
-    size_t num_components, max_component;
-    CountComp(olapondb, label, max_component, num_components);
-    printf("max_component = %ld\n", max_component);
-    printf("num_components = %ld\n", num_components);
+    auto all_vertices = olapondb.AllocVertexSubset();
+    all_vertices.Fill();
+    size_t max_distance_vi = olapondb.ProcessVertexActive<size_t>(
+        [&](size_t vi) {
+            if (distance[vi] > 1e10) {
+                distance[vi] = -1;
+            }
+            return (size_t)vi;
+        },
+        all_vertices, 0,
+        [&](size_t a, size_t b) {
+            return (distance[a] > distance[b] ||
+                    (distance[a] == distance[b] &&
+                     olapondb.OriginalVid(a) < olapondb.OriginalVid(b)))
+                       ? a
+                       : b;
+        });
+    printf("max distance is: distance[%ld]=%lf\n", olapondb.OriginalVid(max_distance_vi),
+           distance[max_distance_vi]);
     auto output_cost = get_time() - start_time;
 
+    auto vit = txn.GetVertexIterator(olapondb.OriginalVid(max_distance_vi), false);
+    auto vit_label = vit.GetLabel();
+    auto primary_field = txn.GetVertexPrimaryField(vit_label);
+    auto field_data = vit.GetField(primary_field);
     // return
     {
         json output;
-        output["num_components"] = num_components;
-        output["max_component"] = max_component;
+        output["max_distance_vid"] = olapondb.OriginalVid(max_distance_vi);
+        output["max_distance_label"] = vit_label;
+        output["max_distance_primaryfield"] = primary_field;
+        output["max_distance_fielddata"] = field_data.ToString();
+        output["max_distance_val"] = distance[max_distance_vi];
         output["num_vertices"] = olapondb.NumVertices();
         output["num_edges"] = olapondb.NumEdges();
         output["prepare_cost"] = prepare_cost;
