+++ procedures/custom_cpp/wcc_procedure.cpp
@@ -1,4 +1,4 @@
-/**
+﻿/**
  * Copyright 2022 AntGroup CO., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,11 +14,31 @@
 
 #include "lgraph/olap_on_db.h"
 #include "tools/json.hpp"
-#include "./algo.h"
+#include "../algo_cpp/algo.h"
 
 using namespace lgraph_api;
 using namespace lgraph_api::olap;
 using json = nlohmann::json;
+
+void CountComp(OlapBase<Empty>& graph, ParallelVector<size_t>& label, size_t& max, size_t& num) {
+    ParallelVector<size_t> cnt = graph.AllocVertexArray<size_t>();
+    cnt.Fill(0);
+    graph.ProcessVertexInRange<size_t>(
+        [&](size_t v) {
+            if (graph.OutDegree(v) == 0) return 0;
+            size_t v_label = label[v];
+            write_add(&cnt[v_label], (size_t)1);
+            return 0;
+        },
+        0, label.Size());
+    max = 0;
+    num = graph.ProcessVertexInRange<size_t>(
+        [&](size_t v) {
+            if (max < cnt[v]) write_max(&max, cnt[v]);
+            return (graph.OutDegree(v) > 0 && cnt[v] > 0) ? 1 : 0;
+        },
+        0, label.Size());
+}
 
 /**
  * Processes a request on a GraphDB.
@@ -28,11 +48,12 @@ using json = nlohmann::json;
  *        The request should contain the following parameters:
  *        - "vertex_label_filter": Filter vertex sets based on vertex labels
  *        - "edge_label_filter": Filter edge sets based on edge labels
- *        - "lcc_value": Vertex field name to be written back into the database.
- *        - "output_file": Lcc value to be written to the file.
+ *        - "wcc_value": Vertex field name to be written back into the database.
+ *        - "output_file": Wcc value to be written to the file.
  * @param response The output response in JSON format.
  *        The response will contain the following parameters:
- *        - "average_lcc": The average lcc of the graph.
+ *        - "num_components": The number of components in the graph.
+ *        - "max_component": The size of the largest component.
  *        - "num_vertices": The number of vertices in the graph.
  *        - "num_edges": The number of edges in the graph.
  *        - "prepare_cost": The time cost of preparing the graph data.
@@ -47,62 +68,44 @@ extern "C" bool Process(GraphDB& db, const std::string& request, std::string& re
 
     // prepare
     start_time = get_time();
-    std::string vertex_label_filter = "";
-    std::string edge_label_filter = "";
-    std::string lcc_value = "";
     std::string output_file = "";
     std::cout << "Input: " << request << std::endl;
     try {
         json input = json::parse(request);
-        parse_from_json(vertex_label_filter, "vertex_label_filter", input);
-        parse_from_json(edge_label_filter, "edge_label_filter", input);
-        parse_from_json(lcc_value, "lcc_value", input);
         parse_from_json(output_file, "output_file", input);
     } catch (std::exception& e) {
         response = "json parse error: " + std::string(e.what());
         std::cout << response << std::endl;
         return false;
     }
-    auto txn = db.CreateReadTxn();
-    std::function<bool(VertexIterator &)> vertex_filter = nullptr;
-    std::function<bool(OutEdgeIterator &, Empty &)> edge_filter = nullptr;
 
-    if (!vertex_label_filter.empty()) {
-        vertex_filter = [&vertex_label_filter](VertexIterator& vit) {
-            return vit.GetLabel() == vertex_label_filter;
-        };
-    }
-    if (!edge_label_filter.empty()) {
-        edge_filter = [&edge_label_filter](OutEdgeIterator& eit, Empty& edata) {
-            return eit.GetLabel() == edge_label_filter;
-        };
-    }
-    OlapOnDB<Empty> olapondb(db, txn, SNAPSHOT_PARALLEL | SNAPSHOT_UNDIRECTED,
-            vertex_filter, edge_filter);
+    auto txn = db.CreateReadTxn();
+    OlapOnDB<Empty> olapondb(db, txn, SNAPSHOT_PARALLEL | SNAPSHOT_UNDIRECTED);
     auto prepare_cost = get_time() - start_time;
 
     // core
     start_time = get_time();
-    auto score = olapondb.AllocVertexArray<double>();
-    double average_lcc = LCCCore(olapondb, score);
+    ParallelVector<size_t> label = olapondb.AllocVertexArray<size_t>();
+    WCCCore(olapondb, label);
     auto core_cost = get_time() - start_time;
 
     // output
     start_time = get_time();
     if (output_file != "") {
-        olapondb.WriteToFile<double>(true, score, output_file);
+        olapondb.WriteToFile<size_t>(true, label, output_file);
     }
-    txn.Commit();
 
-    if (lcc_value != "") {
-        olapondb.WriteToGraphDB<double>(score, lcc_value);
-    }
+    size_t num_components, max_component;
+    CountComp(olapondb, label, max_component, num_components);
+    printf("max_component = %ld\n", max_component);
+    printf("num_components = %ld\n", num_components);
     auto output_cost = get_time() - start_time;
 
     // return
     {
         json output;
-        output["average_lcc"] = average_lcc;
+        output["num_components"] = num_components;
+        output["max_component"] = max_component;
         output["num_vertices"] = olapondb.NumVertices();
         output["num_edges"] = olapondb.NumEdges();
         output["prepare_cost"] = prepare_cost;
@@ -110,6 +113,6 @@ extern "C" bool Process(GraphDB& db, const std::string& request, std::string& re
         output["output_cost"] = output_cost;
         output["total_cost"] = prepare_cost + core_cost + output_cost;
         response = output.dump();
-        return true;
     }
+    return true;
 }
